requirements.txt
tweepy
black
python-dotenv